move subscribeToTablePreference to FshTable component also cleanup userroles
