+++ src/Client/Pages/Identity/Users/UserRoles.razor.cs
@@ -10,25 +10,21 @@ namespace FSH.BlazorWebAssembly.Client.Pages.Identity.Users;
 
 public partial class UserRoles
 {
+    [Parameter]
+    public string? Id { get; set; }
     [CascadingParameter]
     protected Task<AuthenticationState> AuthState { get; set; } = default!;
     [Inject]
     protected IAuthorizationService AuthService { get; set; } = default!;
-    [Parameter]
-    public string? Id { get; set; }
-    [Parameter]
-    public string? Title { get; set; }
-    [Parameter]
-    public string? Description { get; set; }
     [Inject]
     protected IUsersClient UsersClient { get; set; } = default!;
 
-    public List<UserRoleDto> UserRolesList { get; set; } = new();
+    private List<UserRoleDto> _userRolesList = default!;
+
+    private string _title = string.Empty;
+    private string _description = string.Empty;
 
     private string _searchString = string.Empty;
-    private bool _dense;
-    private bool _striped = true;
-    private bool _bordered;
 
     private bool _canEditUsers;
     private bool _canSearchRoles;
@@ -44,14 +40,14 @@ public partial class UserRoles
                 () => UsersClient.GetByIdAsync(Id), Snackbar)
             is UserDetailsDto user)
         {
-            Title = $"{user.FirstName} {user.LastName}";
-            Description = string.Format(L["Manage {0} {1}'s Roles"], user.FirstName, user.LastName);
+            _title = $"{user.FirstName} {user.LastName}";
+            _description = string.Format(L["Manage {0} {1}'s Roles"], user.FirstName, user.LastName);
 
             if (await ApiHelper.ExecuteCallGuardedAsync(
                     () => UsersClient.GetRolesAsync(user.Id.ToString()), Snackbar)
                 is ICollection<UserRoleDto> response)
             {
-                UserRolesList = response.ToList();
+                _userRolesList = response.ToList();
             }
         }
 
@@ -62,7 +58,7 @@ public partial class UserRoles
     {
         var request = new UserRolesRequest()
         {
-            UserRoles = UserRolesList
+            UserRoles = _userRolesList
         };
 
         if (await ApiHelper.ExecuteCallGuardedAsync(
